NDEV-3698 rafactoring tests. Add all methods to TracerClient, extract transaction send logic into fixtures
